+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2
  - use the templating maven plugin to indicate the version of project
  - worker returns void value
+ - create a new class: MessageSenderHelper to send easily a message in as2
 
 ## 1.1 (2015-07-06)
  - add loggers